feat: integrate FeedSettingsModal into Articles and Videos pages
- Imported FeedSettingsModal component into Articles and Videos pages to enhance user customization options for feed preferences.
- Removed the inline FeedSettingsModal implementation from Articles page, streamlining the code structure.
- Updated Videos page to include FeedSettingsModal, allowing users to manage their feed settings directly from the video browsing experience.

These changes collectively improve the user experience by providing a consistent and accessible way to manage feed settings across both Articles and Videos pages.
