+++ src/components/common/FeedSettingsModal/FeedSettingsModal.module.css
@@ -1,0 +1,99 @@
+.modalOverlay {
+  @apply fixed inset-0 z-50
+         bg-black/50 backdrop-blur-sm
+         flex items-center justify-center;
+}
+
+.modal {
+  @apply w-[900px] max-w-[90vw]
+         bg-[#0A0A0A] rounded-2xl
+         border border-white/10
+         overflow-hidden;
+}
+
+.modalHeader {
+  @apply flex items-center justify-between
+         px-6 py-4
+         border-b border-white/10;
+}
+
+.modalTitle {
+  @apply flex items-center gap-3;
+}
+
+.modalTitle h2 {
+  @apply text-lg font-semibold text-white;
+}
+
+.modalTitleIcon {
+  @apply w-5 h-5 text-white/70;
+}
+
+.modalActions {
+  @apply flex items-center gap-2;
+}
+
+.saveButton {
+  @apply px-4 py-2 rounded-lg
+         bg-primary-purple text-white
+         hover:bg-primary-purple/90
+         transition-colors duration-200;
+}
+
+.closeButton {
+  @apply px-4 py-2 rounded-lg
+         text-white/70
+         hover:bg-white/5
+         transition-colors duration-200;
+}
+
+.modalLayout {
+  @apply flex h-[600px];
+}
+
+.sideMenu {
+  @apply w-64 p-4
+         border-r border-white/10
+         space-y-1;
+}
+
+.menuItem {
+  @apply w-full flex items-center gap-3
+         px-4 py-3 rounded-lg
+         text-white/70
+         transition-all duration-200
+         hover:bg-white/5;
+}
+
+.menuItem.active {
+  @apply bg-white/5 text-white;
+}
+
+.menuIcon {
+  @apply w-5 h-5;
+}
+
+.modalContent {
+  @apply flex-1 p-6 overflow-y-auto;
+}
+
+.feedNameSection {
+  @apply space-y-4;
+}
+
+.feedNameSection h3 {
+  @apply text-lg font-medium text-white;
+}
+
+.sectionDescription {
+  @apply text-sm text-white/70;
+}
+
+.feedNameInput {
+  @apply w-full px-4 py-2 rounded-lg
+         bg-white/5
+         border border-white/10
+         text-white
+         focus:outline-none focus:border-primary-purple
+         transition-colors duration-200;
+}
